moved version from github to util
