+++ pkg/util/githubutil/githubutil_test.go
@@ -13,8 +13,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-const testVersion = "v0.1.20"
-
 func TestGetLatestVersion(t *testing.T) {
 	t.Run("Get latest release with wrong url", func(t *testing.T) {
 		_, err := GetLatestVersion("fl")
@@ -123,38 +121,4 @@ func TestGetUpgradeMessage(t *testing.T) {
 	message, err = GetUpgradeMessage(version, linux)
 	assert.Nil(t, err)
 	assert.Equal(t, 157, len(message))
-}
-
-func TestIsVersionGreaterThan(t *testing.T) {
-	t.Run("Compare FlyteCTL version when upgrade available", func(t *testing.T) {
-		_, err := IsVersionGreaterThan("v1.1.21", testVersion)
-		assert.Nil(t, err)
-	})
-	t.Run("Compare FlyteCTL version greater then", func(t *testing.T) {
-		ok, err := IsVersionGreaterThan("v1.1.21", testVersion)
-		assert.Nil(t, err)
-		assert.Equal(t, true, ok)
-	})
-	t.Run("Compare FlyteCTL version greater then for equal value", func(t *testing.T) {
-		ok, err := IsVersionGreaterThan(testVersion, testVersion)
-		assert.Nil(t, err)
-		assert.Equal(t, false, ok)
-	})
-	t.Run("Compare FlyteCTL version smaller then", func(t *testing.T) {
-		ok, err := IsVersionGreaterThan("v0.1.19", testVersion)
-		assert.Nil(t, err)
-		assert.Equal(t, false, ok)
-	})
-	t.Run("Compare FlyteCTL version", func(t *testing.T) {
-		_, err := IsVersionGreaterThan(testVersion, testVersion)
-		assert.Nil(t, err)
-	})
-	t.Run("Error in compare FlyteCTL version", func(t *testing.T) {
-		_, err := IsVersionGreaterThan("vvvvvvvv", testVersion)
-		assert.NotNil(t, err)
-	})
-	t.Run("Error in compare FlyteCTL version", func(t *testing.T) {
-		_, err := IsVersionGreaterThan(testVersion, "vvvvvvvv")
-		assert.NotNil(t, err)
-	})
 }
