+++ pkg/util/util.go
@@ -8,6 +8,8 @@ import (
 	"os"
 	"strings"
 
+	hversion "github.com/hashicorp/go-version"
+
 	"github.com/enescakir/emoji"
 	"github.com/flyteorg/flytectl/pkg/configutil"
 	"github.com/flyteorg/flytectl/pkg/docker"
@@ -73,4 +75,17 @@ func SendRequest(method, url string, option io.Reader) (*http.Response, error) {
 		return nil, fmt.Errorf("someting goes wrong while sending request to %s. Got status code %v", url, response.StatusCode)
 	}
 	return response, nil
+}
+
+// IsVersionGreaterThan check version if it's greater then other
+func IsVersionGreaterThan(version1, version2 string) (bool, error) {
+	semanticVersion1, err := hversion.NewVersion(version1)
+	if err != nil {
+		return false, err
+	}
+	semanticVersion2, err := hversion.NewVersion(version2)
+	if err != nil {
+		return false, err
+	}
+	return semanticVersion1.GreaterThan(semanticVersion2), nil
 }
